Cardano.Wallet.Shelley.Api.Server

Description

API handlers and server using the underlying wallet layer to provide endpoints reachable through HTTP.

Documentation

server :: forall t n. (Buildable (ErrValidateSelection t), PaymentAddress n IcarusKey, PaymentAddress n ByronKey) => ApiLayer (RndState n) t ByronKey -> ApiLayer (SeqState n IcarusKey) t IcarusKey -> ApiLayer (SeqState n ShelleyKey) t ShelleyKey -> NtpClient -> Server (Api n)

A diminished servant server to serve Byron wallets only.