Cardano.Wallet.Shelley.Compatibility

Description

Conversion functions and static chain settings for Shelley.

Documentation

data Shelley

type ShelleyBlock = ShelleyBlock TPraosStandardCrypto

type NodeVersionData = (NodeToClientVersionData, CodecCBORTerm Text NodeToClientVersionData)

data TPraosStandardCrypto

Chain Parameters

mainnetVersionData :: NodeVersionData

Settings for configuring a MainNet network client

testnetVersionData :: ProtocolMagic -> NodeVersionData

Settings for configuring a TestNet network client

mainnetBlockchainParameters :: GenesisBlockParameters

Genesis

emptyGenesis :: BlockchainParameters -> Block

genesisTip :: Tip (ShelleyBlock TPraosStandardCrypto)

Conversions

toShelleyHash :: Hash "BlockHeader" -> ShelleyHash c

toEpochSize :: EpochLength -> EpochSize

toGenTx :: HasCallStack => SealedTx -> GenTx ShelleyBlock

SealedTx are the result of rightfully constructed shelely transactions so, it is relatively safe to unserialize them from CBOR.

toPoint :: Hash "Genesis" -> EpochLength -> BlockHeader -> Point ShelleyBlock

toSlotNo :: EpochLength -> SlotId -> SlotNo

fromBlockNo :: BlockNo -> Quantity "block" Word32

fromShelleyBlock :: Hash "Genesis" -> EpochLength -> ShelleyBlock -> Block

fromShelleyHash :: ShelleyHash c -> Hash "BlockHeader"

fromPrevHash :: Hash "BlockHeader" -> PrevHash TPraosStandardCrypto -> Hash "BlockHeader"

fromChainHash :: Hash "Genesis" -> ChainHash ShelleyBlock -> Hash "BlockHeader"

fromGenesisData :: ShelleyGenesis TPraosStandardCrypto -> (GenesisBlockParameters, [TxOut])

Convert genesis data into blockchain params and an initial set of UTxO

fromNetworkMagic :: NetworkMagic -> ProtocolMagic

fromSlotNo :: EpochLength -> SlotNo -> SlotId

fromTip :: Hash "Genesis" -> EpochLength -> Tip ShelleyBlock -> BlockHeader

fromPParams :: PParams -> TxParameters