Cardano.Wallet.Shelley.Network

Top-Level (Stateless) Interface

This is a first attempt at using the ChainSync mini protocol in a stateless manner. Note that this is a bit unfortunate as the the mini-protocols are intrinsically stateful but our network layer is currently stateless. So here, I've turned the mini-protocols into a stateless API, although, we might want to do it the other way around and make our network stateful :)

Instead of passing around the block header hash to get the next blocks, we could let the network layer keep track of where the pointer is. We'd need a way to resolve an initial intersection of course when spinning up a network layer.

Having said that, it's quite an important design change that will really pay off when dealing with rollback. So, in this first iteration, we may simply use the mini-protocols in a stateless fashion and review our network layer design to work statefully (including for Jormungandr) when we start working on rollbacks.

data NetworkLayer m

A stateless interface for dealing with the Haskell NetworkClient

See newNetworkClient and connectClient methods to construct a NetworkLayer:

let nodeId = CoreNodeId 0 let addr = localSocketAddrInfo (localSocketFilePath nodeId) let params = ChainParameters { epochSlots = EpochSlots 21600 , protocolMagic = ProtocolMagicId 1097911063 } (client, network) <- newNetworkClient nullTracer params

- Launch the stateful connection in a separate thread void $ forkIO $ connectClient client dummyNodeToClientVersion addr

- Use the network layer to send requests to the node response <- runExceptT $ nextBlocks network genesisPoint point <- networkTip network response' <- runExceptT $ nextBlocks network point

Constructors

NetworkLayer

Fields

nextBlocks :: Point -> ExceptT ErrNextBlocks m [Block]

networkTip :: m Point

data ErrNextBlocks

What is considered errors in the NetworkLayer

Constructors

ErrNextBlocksNoIntersection

ErrNextBlocksRollBack

data ChainParameters

Constructors

ChainParameters

Fields

epochSlots :: EpochSlots

Number of slots per epoch.

protocolMagic :: ProtocolMagicId

Protocol magic (e.g. mainnet=764824073, testnet=1097911063)

Re-Export

newtype EpochSlots

Constructors

EpochSlots

Fields

unEpochSlots :: Word64

newtype ProtocolMagicId

Constructors

ProtocolMagicId

Fields

unProtocolMagicId :: Word32

Constructors

newNetworkClient :: forall m. (MonadThrow m, MonadST m, MonadTimer m) => Tracer m String -> ChainParameters -> m (NetworkClient m, NetworkLayer m)

Construct a new network client handling ChainSync and LocalTxSubmission mini-protocols.

connectClient :: forall vData. vData ~ NodeToClientVersionData => NetworkClient IO -> (vData, CodecCBORTerm Text vData) -> AddrInfo -> IO ()

dummyNodeToClientVersion :: (NodeToClientVersionData, CodecCBORTerm Text NodeToClientVersionData)

A dummy network magic for a local cluster. When connecting to mainnet or testnet, this should match the underlying network's configuration.

Transport Helpers

localSocketAddrInfo :: FilePath -> AddrInfo

Construct an AddrInfo from a FilePath. This caracterises a unix domain socket than serves as transport for the mini-protocols.

localSocketFilePath :: CoreNodeId -> FilePath

Sockets are named after a particular convention using a core node id.